backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pypdf2>=3.0.0
groq>=0.4.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
numpy>=1.26.4
sentence-transformers>=2.2.0
edge-tts>=6.2.0
websockets>=12.0
pyttsx3>=2.90
Pillow>=10.0.0
requests>=2.31.0
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0
firebase-admin>=6.0.0
chromadb>=0.4.0
opencv-python-headless
scipy
pandas
tqdm
librosa
imageio
imageio-ffmpeg